+++ src/main/java/br/com/will/classes/virtualthread/infrastructure/gateway/payment/PaymentGatewayImpl.java
@@ -3,6 +3,8 @@ package br.com.will.classes.virtualthread.infrastructure.gateway.payment;
 import br.com.will.classes.virtualthread.port.PaymentGateway;
 import org.springframework.stereotype.Component;
 
+import java.math.BigDecimal;
+
 @Component
 public class PaymentGatewayImpl implements PaymentGateway {
 
@@ -13,7 +15,7 @@ public class PaymentGatewayImpl implements PaymentGateway {
     }
 
     @Override
-    public String charge(String orderId, double amount) {
+    public String charge(String orderId, BigDecimal amount) {
         return client.pay(new PaymentRequest(orderId, amount));
     }
