+++ src/main/java/br/com/will/classes/virtualthread/infrastructure/gateway/payment/PaymentRequest.java
@@ -1,4 +1,6 @@
 package br.com.will.classes.virtualthread.infrastructure.gateway.payment;
 
-record PaymentRequest(String orderId, double amount) {
+import java.math.BigDecimal;
+
+record PaymentRequest(String orderId, BigDecimal amount) {
 }
