+++ src/main/java/br/com/will/classes/virtualthread/port/FraudChecker.java
@@ -1,7 +1,9 @@
 package br.com.will.classes.virtualthread.port;
 
+import java.math.BigDecimal;
+
 public interface FraudChecker {
 
-    boolean isFraud(String orderId, double amount);
+    boolean isFraud(String orderId, BigDecimal amount);
 
 }
