+++ src/main/java/br/com/will/classes/virtualthread/port/PaymentGateway.java
@@ -1,7 +1,9 @@
 package br.com.will.classes.virtualthread.port;
 
+import java.math.BigDecimal;
+
 public interface PaymentGateway {
 
-    String charge(String orderId, double amount);
+    String charge(String orderId, BigDecimal amount);
 
 }
